+++ CHANGELOG.md
@@ -28,6 +28,7 @@
   respectively - fixes [Issue #249](https://github.com/PlagueHO/CosmosDB/issues/249).
 - Changed `Remove-CosmosDbAccount` to prevent second confirmation prompt
   when removing account.
+- Enabled `*-CosmosDbAccount` tests to run in AppVeyor.
 
 ## 2.1.15.237